Moves `<GameLoopScreen>` to its own folder
