+++ src/web-ui/screens/game-loop/price.css
