+++ src/web-ui/screens/game-loop/priceFormatter.ts
@@ -1,0 +1,7 @@
+const userLocale = navigator.language || 'en-US';
+const priceFormatter = new Intl.NumberFormat(userLocale, {
+  style: 'currency',
+  currency: 'USD',
+});
+
+export default priceFormatter;
